Read Me
Create a class named BloodData that includes fields that hold a blood type (the four blood types: O, A, B, and AB) and an RH factor (the factors are + or -).
Create a default constructor that sets the fields to "O" and "+", and an overloaded constructor that requires values for both fields. 
Include get and set methods for each field. save the file as BloodData.java. 
Create an application named TestBloodData that demonstrates each method works correctly. Save the application as TestBloodData.java.



Create a class named Patient that includes an ID number, age, and BloodData.
Provide a default constructor that sets the ID number to 0, the age to 0, and the BloodData to "O", and "+". 
Create an overloaded constructor that provides values for each field. 
Also provide get methods for each field. Save the file as Patient.java.
Create an application that demonstrates that each methods works correctly , and save it as TestPatient.java.